add and move files in order for theme to work
