+++ src/@lekoarts/gatsby-theme-cara/sections/contact.mdx
@@ -1,0 +1,3 @@
+## Get in touch
+
+Visit [my website](https://www.lekoarts.de?utm_source=cara&utm_medium=Starter) or find me on other platforms: [Twitter](https://www.twitter.com/lekoarts_de) & [LinkedIn](https://www.linkedin.com/in/lennart-joergens/)
