+++ src/@lekoarts/gatsby-theme-cara/sections/intro.mdx
@@ -1,0 +1,3 @@
+# Hi, I'm Andy
+
+Template by [lekoarts.de](https://www.lekoarts.de?utm_source=cara&utm_medium=Starter) - I write about JavaScript and React there!
