+++ src/pages/archive/index.tsx
@@ -1,6 +1,6 @@
 import * as React from "react";
-import Layout from "../components/layout";
-import Seo from "../components/seo";
+import Layout from "../../components/layout";
+import Seo from "../../components/seo";
 
 const IndexPage = () => {
   return (
